🎨 Move pricing options generation to utils
