+++ constant/index.ts
@@ -1,30 +1,9 @@
-import pricingData from '~/assets/data/pricing_option.json'
-
 export const IPFS_VIEW_GATEWAY_URL = 'https://ipfs.io/ipfs'
 
 export const MINIMAL_PRICE = 0.99
 export const DEFAULT_PRICE = 4.99
 export const DEFAULT_PRICE_STRING = '4.99'
 export const MAXIMUM_PRICE = 99.99
-
-function generatePricingOptions () {
-  const options = [{ label: '免費', value: '0' }] // Use string '0' to avoid falsy value issues in UI components
-
-  Object.entries(pricingData).forEach(([_key, data]) => {
-    const usdValue = data.USD
-    const hkdValue = data.HKD
-    const twdValue = data.TWD
-
-    const twdFormatted = `NT$${twdValue.toLocaleString()}`
-
-    options.push({
-      label: `${usdValue.toFixed(2)} (HK$${hkdValue}, ${twdFormatted})`,
-      value: usdValue.toString()
-    })
-  })
-
-  return options
-}
 
 export const USD_PRICING_OPTIONS = generatePricingOptions()
